+++ config/tree/tcp_config.go
@@ -2,7 +2,7 @@ package tree
 
 type TcpConfig struct {
 	BindIp           string `validate:"required,ip"`
-	BindPort         uint   `validate:"required"`
+	BindPort         uint16 `validate:"required"`
 	BindZone         string
 	ClientBufferSize uint `validate:"required"`
 	HostBufferSize   uint `validate:"required"`
