+++ config/tree/udp_config.go
@@ -2,7 +2,7 @@ package tree
 
 type UdpConfig struct {
 	BindIp     string `validate:"required,ip"`
-	BindPort   uint   `validate:"required"`
+	BindPort   uint16 `validate:"required"`
 	BindZone   string
 	BufferSize uint `validate:"required"`
 }
